correção de bugs nos testes

+++ src/main/java/app/service/VendaService.java
@@ -42,6 +42,9 @@ public class VendaService {
 	
 	public String delete (long id) {
 		
+		if(id < 0)
+			throw new RuntimeException("Objeto não pode estar nulo");
+		
 		this.vendaRepository.deleteById(id);
 		return "Venda deletado com sucesso!";
 	}
